requirements.txt
Pillow>=9.0.0
django
python-dotenv
djangorestframework
drf-yasg
django-cors-headers